Give the Royal Jelly suitably royal case
This makes the unique have the name "the Royal Jelly" instead of "the
royal jelly", so that it's not the only unique with an all lower-case
name. RC file options are case-insensitive unless explicitly made
otherwise through use of regexp, so this shouldn't break any options,
but any lua code in your RC that does an explicit string matching to
"the royal jelly" will need to be updated to look for the new case.

This commit also cleans up usage of the name in all des files and
comments to be consistent with the new case.