Don't crash when spectral Gyre kills itself on spines.
To reproduce the crash, create a hell sentinel (preferably paralysed or
otherwise immobile and with enough HP to survive stabs), wield Gyre,
cast Spectral Weapon, bring the SW down to 1 hit point, and attack the
sentinel.  After a few tries, the SW will be struck by spines, causing a
crash (or possibly a freeze, depening on the details of the heap
corruption).  The problem is that our unwind_var stores a reference to
the value of the ARTEFACT_NAME_KEY property, but resetting the item
or even adding a new property can invalidate that reference.

Use an unwinder rather than an unwind_var to restore the weapon's name
only if it is still the same weapon.