Re-correct weights for random tiles.
The fix in 0.18-a0-456-gaa2475c handles the case of identical weights
(assuming the base tile has the smallest index, which is always the
case), but otherwise uses the wrong weights.

For example, if the weights are  A:2, B:3, C:1, then total = 6 and we
roll a "random" number from 0 to 5:

Roll    Old   aa2475c   Now
 0       A       A       A
 1       A       B       A
 2       A       B       B
 3       B       C       B
 4       B       C       B
 5       B       C       C

If the weights are  A:2, B:2, then total = 4:
Roll    Old   aa2475c   Now
 0       A       A       A
 1       A       B       A
 2       A       B       B
 3       B    base idx   B

Ideally both this and random_choose_weighted would be using the same
code, since the logic is essentially identical (RCW adds to the number
being compared against rather than subtracting from the roll, but that
amounts to the same thing).