Give hydra form a fixed number of heads when cast
Letting it grow more heads in combat was a bad gameplay mechanic, but removing
that resulted in it being unecessarily ineffective against eg humanoids with
chopping weapons. Just giving it a fixed number of heads when cast makes it
simpler (at the cost of making it more unlike monster hydras, but again, it was
already really nothing like them).