Remove most sources of Zin Recite interruption
The list of actions that interrupted recite was very long and awkward, and it
wasn't a very necessary restriction. Flavour-wise, if you can carry on casting,
quaffing and reading while singing a song of slaying it's just as believable
that you could do so while reciting!

Losing a large amount of HP also no longer interrupts recite - this was just a
holdover from uninterruptible recite.