Adjust Device Surge into a slimify-like one-shot duration (MarvinPA).
The old interface was clunky; this is considerably cleaner and also
allows for the surge to not dissipate if you're using a device to which
it doesn't apply.

I've flagged the ability as instant so that it takes the same amount of
time to do a device surge as previously, i.e. the time it takes to evoke
the device. This may not turn out to be necessary or desirable, but it's
there for now.