Mark stone stairs as known if you have seen both ends of them.
Previously it was generally a good idea to take each upstairs when exploring
a new level, since this gave the following benefits:
* you might need to do this to access a disconnected part of the previous
  level (often a vault)
* you would know what the terrain was like on the other side of the stairs
  in case you later wanted to retreat up the stairs while pursued by monsters
* interlevel travel would be able to use those stairs to move between levels
  more efficiently (this is especially important for the orbrun)

But if you had already fully explored the previous level then taking an
upstairs to it was very unlikely to be interesting at all, despite still
giving benefits 2 and 3 above. Manually doing <> on each unknown upstairs was
an uninteresting task (and one that autoexplore does not do for you).

This commit effectively gives benefits 2 and 3 to you for free by marking
stone stairs as known (and letting interlevel travel know about them) as soon
as you have seen both sides of them. (This information was actually already
leaked by player lua asking for the name of the stair feature.) This means
that if an upstairs is marked as unknown then you know for sure that there
is something new on the other side of it that you should go explore.

This commit does lose the information of which stairs a player has actually
traversed. If it turns out that players really want the game to keep track
of this for them, I'd suggest adding it as a flag in stair_info so that it
is logically separate from knowing the destination of the stair. Then an
option can be added to emphasise stairs based on whether the stair has been
traversed rather than on whether the destination of the stair is known. I'm
not adding a new option until it is clear that such behavior is desired,
though...