Pakellas gifts a wand once at 2*.
This is intended to improve P's utility in the early game, where it can
be difficult to do anything in the absence of a wand.

The rod gift at 4* is preserved, and the 2* gift does not impact further
gifts.

The wand is guaranteed not to be heal wounds, teleportation, or hasting
(P can gift those later if you're sufficiently devout, of course).