Slow rate of recharging of mostly full stacks of wands.
The idea behind this is that your wands in a stack are all at roughly
the same charge level, so the maximum number of charges you can get out
of a recharging scroll is the maximum number of charges you could get
by charging any one of those wands in the stack.

Basically, the more charged the stack is, the less effective recharging
is on it - just like for individual wands.