Fix a hang when deleting clouds.
Invalidating iterators to the current element is a problem, because the
for loop's iterator points to that element and still needs to be
incremented.

I couldn't reproduce the hang on the servers, but locally it happened
whenever I placed a cloud and changed levels.  Could perhaps in some
situations not hang, but leave clouds over on the new level, causing a
crash?