fix a novelty vault to work better with squarelos
see !bug 10120
https://crawl.develz.org/mantis/view.php?id=10120

The old silent_spectre radius under circlelos covered a diameter of 25
squares.  In squarelos, the "diameter" is only 21.

This meant a few monsters were not properly silenced per the intention
of the vault.

One solution would have been to increase silence_radius to 12, but this
seems like a poor choice just to make a single novelty vault work
better.

Instead, the proposed fix makes a couple cells on the East side only 2
squares (just like some other existing squares).

Additionally, the silent_spectre is moved one space EAST to help cover
the new smaller cells.

Loot and number of mosnters remains unchanged. However, the ancient
lich, 6, is now BEHIND the diggable rock wall at the south and
completely outside of the silenced area.

If this is not desired, it could easily be moved to the 4 just above the
diggable walls.