Redo penalties for amulets of regen and dismissal
The old penalties gave strategic penalties for tactical costs, which
seemed disproportional and inappropriate.

The new swap deterrent for amulets of regeneration (courtesy of Haifisch
via archaeo, I believe) is that the amulet won't regenerate you at all
until it attunes to your physiology, which happens as soon as you hit
full health. This means that you cannot use it for emergency healing, but
there's no frustrating penalty for swapping it strategically.

The new swap deterrent for amulets of dismissal is a brief (3-6 turns)
spell of vertigo when you put it on. Vertigo affects your EV, combat
accuracy, and spell success chances. This makes it a bad thing to pop on
in the middle of a fight, but is not a meaningful penalty outside of
combat. The status lasts quite briefly, meaning that resting after
putting on the amulet isn't particularly necessary.