Transform the amulet of stasis into the amulet of reflection
The amulet of stasis was an all-or-nothing resistance to a variety of
effects, many of which were good effects. As such, it was only used as a
swap, generally against paralysis, or sometimes to suppress mutations.
Because its use completely protected against the statuses involved, it
wasn't a particularly interesting way to deal with those statuses. A few
vaults still have permanent teleportation traps where stasis could be
useful, but for those vaults Apportation still exists. Permanent tele
traps can still spawn blocking one or both lungs of Zot:5, but these
aren't hard barriers, since trying them often enough will eventually
teleport one past them, and blinking past them is still an option.

The amulet of reflection gives you a variable amount of SH and the
reflection status, allowing you to reflect projectiles shot at you as if
you had a shield of reflection, whether or not you do. The egos that compete
with reflection on shields often make it the worst choice; making it
available on an amulet that also gives bonus SH makes it a significantly
more compelling option, giving players another source for an interesting
attribute.

This commit also adds in SH as an artefact property and switches the amulet
of Shielding into an amulet of reflection that gives +8 shielding; this makes
the amulet a parallel to the rings of Shaolin and Robustness.

Oh, and reflection code could probably stand to be refactored,
but I'm going to push that off to another day.