Fix the targetter sometimes claiming the last square was hit twice.
This happened when the player's cursor ended the beam in the wall.
affect_cell() called hit_wall(), which, if it the beam couldn't affect the
wall, regressed the ray until it was back in open space and set
extra_range_used to signal that the beam was over and the last space
wasn't supposed to actually be used twice. Unfortunately, the addition to
path_taken was after affect_cell but before the check to break out of the
loop if the beam had finished. This previously didn't cause issues,
because path_taken didn't add a square if it had just been there, but that
check had to be removed in order to properly show a beam bouncing off a
wall.

The simple solution would be to move the addition to path_taken to before
affect_cell being called. Unfortunately, affect_cell also called bounce(),
and not using the updated ray position would mean dropping the first cell
after the bounce. Therefore, the code to update the position had to be
pulled into the while loop in do_fire().

(cherry picked from commit ce59692f9b4f2f5223d0284142fc1031e8d472b5)