Adjust Xom's random spell list
Remove a few boring spells (Blink, Beastly Appendage, Stoneskin), add some new
ones (Disjunction, Leda's Liquefaction, Summon Mana Viper), and re-order the
list - it's implemented so that later spells are only possible to cast at
higher severity, but there didn't appear to be much logic in the existing
ordering (with lots of summons/transmutations being grouped together, some
particularly high-level spells near the beginning of the list, and so on). The
new ordering is pretty arbitrary but is very roughly by spell level.