Be more careful about creating dummy zombies (Medar)
Zombies should always be created with define_zombie, not define_monster.
Fixes the new HP>0 assert triggering when checking whether a zombie will
anger gods (specifically, in Shadow Creatures).

   !lm acwest crash 30 -log

It would be nice to make define_monster with a zombified monster type
call define_zombie.  However, the problem is that we need a base type,
which is probably missing---and it isn't clear what type we should
use if it is unspecified.