Spread out linesprint high-tier uniques more (bsdbeard)
A bad nsubst meant that, when the map was horizontal, the high-tier
uniques in the middle of the map (Louise, Nikola, Boris, etc) were
bunched up into one line, which might be earlier or later depending on
horizontal flipping.  When the map was vertically oriented, the uniques
were instead staggered across rows, but always at the top of the map.

Now they are distributed randomly (and not all guaranteed to place), as
appears to have been the intent of that nsubst.  To keep difficulty from
going down too much, the number of uniques placed is doubled compared to
the original broken nsubst.

[skip ci]