Replace the amulet of clarity with the amulet of pain
The problems with the amulet of clarity are similar to the ones with the
amulet of stasis: complete immunity to a range of threats and no influence
when those threats are absent. Unlike stasis, the amulet also had no real
downside to make it interesting, encouraging you to carry it around and swap
it in whenever the set of effects it handles are around and you have
insufficient MR. Also unlike stasis, this amulet had weird corner cases
such that it allowed voluntary berserk but not involuntary berserk, but
never allowed ambrosia.

The one place where amulet of clarity really felt useful was on mummies,
since confusion is everywhere and long-lasting, and has a common cure not
available to mummies. In general, finding MR will resolve this; the exception
is mummies worshipping Xom, since Xomfusion can't be resisted normally. I'd
rather address that situation directly than keep this amulet just to handle
this situation.

Meanwhile, the new amulet of pain is an evil amulet that increases damage
dealt to and by the wearer by 25%. If both the attacker and defender have
the amulet, the damage is only increased by 25% total. I expect this
amulet to be strong to characters with great defenses or ranged attacks,
and perhaps speedrunners. Letting this amulet swap freely would encourage
people to use it against monsters that can't fire back and then swap it
when the monster gets into range, so I gave it draining as a swapping
penalty.

This is one of the more experimental designs in the new set of amulets, and
as such I expect it'll get tweaked a few times and it may eventually get
removed or replaced.