Two Pakellas overflow vaults
A couple of vaults themes as experiments conducted by Pakellas in front
of the player. They've been carefully made so they won't break anything
and so they only trigger when the vault is properly visible to the
player. These vaults are in a separate file since there's a fair amount
of accompanying lua.