Sometimes upgrade plain decks for Nem worshipers
Currently, Nemelex worshipers get a card power bonus based on Piety, but
plain decks can NEVER be upgraded past level 0, making this bonus 100%
worthless for plain decks.

While there are more serious criticisms of the current Nemelex
implementation, it seems odd that the God of Decks doesn't provide any
bonus at all to plain decks (besides gifting them a bunch).

This update provides a very small chance for plain decks to be upgraded
to level 1.  This is roughly 1/2 the chance of an ornate deck to be
upgraded (remember that plain decks do not get the +150 bonus to power).