Adjust display of amulet effects on % screen
List faith, harm and reflection alongside the already-displayed spirit and
dismissal. Shuffle SustAt into the first column (mostly just to free up some
space), and only display rMut, Stasis, and Clarity/*Rage when they're actually
active (as was previously the case for -Cast).

It might be reasonable to rethink what's displayed on this screen and where,
for now this just makes the amulets mostly-consistent. Regeneration still isn't
listed (it might be fairly complicated to display well, with so many different
sources of it), and nor is Inaccuracy.