Warn the player if they are tornadoed above deep water / lava.
In watching several instances of players dying in this fashion, I've
noticed that they don't seem to notice that they're in peril. This is
intended to at least bring attention to that fact.

We may need to do more to address this, but let's see if drawing
attention to the issue helps alleviate it in the short term.