Abort the targetter entirely when choosing a bad position (#1531).
Such as the player on a no-self-target spell, after saying [N]o to the
prompt on "Really target yourself?", or out of range.

This is more consistent with other prompts, like for shooting a beam at an
ally. It should also be easier to make it clear what's going on. I won't
deny this may be confusing at first for veterans, but I hope it's easy
enough to get used to.