Place death knights earlier; remove Haunt from them.
They're basically skeletal warriors with spells and a different band of
allies, so placing them considerably later than skeletal warriors
normally appear is boring.

They now place in late D (in which case they don't get a ghoul / flayed
ghost as part of their band), Vaults, and at a lower weight in Crypt
(and not as OODs). Haunt goes away because it would be too powerful
against the player in some of those early placements - and the monsters
they come with generally prove to be a more interesting assortment than
the Haunt wraiths.