Show immunity when targetting an MR-resistible spell (#9391).
Pain on a zombie, for example, will show "not susceptible" rather than an
irrelevant percentage.

As mentioned in comments, this is implemented with an ultra-wrong hack, but
the alternative seems to be massive code duplication. This relates back to
fundamental issues with the place of monster_info in the codebase, which
are beyond the scope of this commit.