Base Spectral Weapon's offensive scaling on spellpower rather than weapon skill
Primarily to simplify it and to make it less effective later on. HD now scales
up to 27 at 100 power, and the damage multiplier scales as if having 12.5
weapon skill under the old system (so at 100 power the damage multiplier is
1.5).

The main effect should be to reduce how well it scales later on, as well as
requiring higher power for it to do so (rather than just continuing to train
weapon skills, which most users would be doing anyway, and also bearing in mind
that it still improves with better base weapons anyway).