Don't badly aim rod of lightning/clouds southeast.
If the player aimed the shotgun or thunderbolt targetter past a grate or
at themselves, the beam would mysteriously point southeast instead---it
turns out, directly away from (0,0).  This happened because, when making
the fallback ray, we passed start and end coordinates to geom::ray's
constructor, which actually expected a start coordinate and a direction.

Also, don't allow self-aiming the shotgun targetter at all (return false
early), since we'd just get a crash as we tried to advance a ray with
direction (0, 0).