Properly round rays with negative coordinates.
We wanted a number to be odd, but it could have been negative, in which
case the mod operator would have given -1 (in C++11 at least) rather
than the expected 1.

The previous commit now prevents us from being called in that situation,
but we should still handle this case, because sometimes we're not in
the map's coordinate system anyway.

Both this commit or the previous one independently fix this crash:
  http://dobrazupa.org/morgue/GreatCaesar/crash-GreatCaesar-20160105-193033.txt
(aiming a cloud targetter north or west when in the vicinity of (0,0)),
but not the more general ray-rounding crash.