Make iron giants throw from the hip.
That is, start the SPELL_THROW beam from the thrower's position as
opposed to the victim's.  That was how Robin's SPELL_THROW_ALLY worked
already, and importantly _choose_throw_target, which served as the
tracer for both spells, worked that way.

Fixes a crash when there was a path from the thrower to the target but
not from the victim to the target: the call to find_ray from
_choose_throw_dest failed, triggering an assert when we tried to advance
the ray:

  http://crawl.akrasiac.org/rawdata/tomatochips/crash-tomatochips-20160104-123826.txt

Also add an explicit assert that we found a ray to trace, so that future
crash messages will credit mon-cast.cc rather than the fearsome and
inscrutable ray.cc.

This change does mean that you can now be thrown to places you could not
see from your old position.  For example, in the following diagram, the
spot labelled * is now a valid destination for the throw:

  #*#.#
  #.#@#
  #.C.#
  #####