Replace ring of invisibility with ring of resist corrosion
This is the first part of a set of commits that will remove the amulet of
resist corrosion, thereby effectively moving resist corrosion from an amulet
to a ring.

The motivation here is to keep all the similar resists (rF+, rC+, rN+, rPois)
together in the ring slot rather than spreading them out. It also means that
we can move another amulet away from being a situational resist swap.