Allow messages to repeat more reliably
"You hear a shout! x10" only worked if that was the first message
of a line; multiple messages on a line were treated as their
concatenation ("You hear a buzzing! You hear a shout!") and so
didn't stack ("You hear a buzzing! You hear a shout! You hear a
shout!").

This is a little ugly, in both code and aesthetics, but should be
an improvement in usability, at least.