Replace the amulet of resist corrosion with the amulet of mana regeneration
The amulet of corrosion is now redundant, so I'm replacing it with something
else. One option is clearly an amulet of invisibility, but I don't actually
like having evokable invis on jewellery -- it's not a great thing to have
on a swappable slot for one thing, and invis is both powerful and has a ton
of sources for another.

Instead, I'm adding an amulet of mana regeneration. It has a warm-up, like
the amulet of regeneration to disincentivize constant swapping. It
increases MP regeneration by the same amount as adding 40 mp would. If it
proves too strong, we can reevaluate.

I think the new amulet is a solid if not particularly flashy item with a
unique enough effect, and I expect it to be useful to many characters.