Improve and ruin and up/downgrade the flavor on Alistair's Intoxication
Alistair's intoxication no longer confuses the player or drains int; instead
it gives Vertigo status. Alistair's also can affect rPois monsters, but at
the same rate that they get poisoned from other sources. Because this is
mostly a buff to the power level, I'm bumping the level to 5. That might
be too much -- it could also be a level 4 spell and add Hexes or
something.

In the process, I made the flavor better (subjectively). The flavor of
destroying your own brain to create poisonous radiation that makes
intelligent monsters drunk never made any sense to me. Now, you destroy
the brains of your foes instead, and you get vertigo from the strange
experience of magically contacting all those brains.

Happily, Alistair's remains a party trick, both in description and book.

This is the result of conversation about the overlap between this spell
and mass confusion; we decided this was the more interesting one to keep,
so I expect we'll remove mass confusion shortly.