Fix jewellery tiles and add a temporary tile for mana regeneration
Move the ego tile for resist corrosion to rings and apply it to rings of
resist corrosion, rename the warding tiles to dismissal. The corrosion
tile needs to be shifted and resized to be consistent with the other
ring tiles, but I've not done this.

This commit also adds a temporary mana regeneration tile as a magenta
palette swap of the regeneration tile. Replace it when a more
appropriate tile is available.