Add Ukayaw as a god that compiles
Ukayaw the Reveler is the implemention of the fast-piety/limit-break/combo
god that's been suggested in a few different forms recently. The god quickly
gains piety while in combat (piety for each instance of dealing damage
and piety for damage dealt), and loses piety incredibly quickly out of battle
down to a baseline value. The granted powers are each exponentially more
powerful than the last so that there's strong incentive to reach higher piety.
Each power should involve elements of positioning and damage while staying
on the themes of ecstatic passion/dance/revelry.