Fix bugs with animated monsters' equipment.
In addition to the ones noted in the removed comment on _equip_undead,
undead versions of Gastronok/Nikola/draconians/etc. are able to use their
peculiar armours.

Undead can now use items that they died with even if you picked them up in
the meantime. If this is undesirable, it should be easy to change.