Fix conditional trove weights
Replace them with vetoes, so that if you don't meet the condition for
a trove, you have to hit a one in x chance. Effective weights upon
failing the condition are the same as before. (except they actually
work now!)

trove_misc_throne_room was newly given weight 15 with the condition
Evocations skill>4, with a 1/3 chance of being chosen if the condition
was failed, so effectively weight 5 in that case, and weight 0 if you
sacrificed Evo to Ru.