Ukayaw *** power: Ukayaw prepares the audience for your solo
Upon reaching ***, Ukayaw briefly paralyzes all enemies on LOS. The duration
is at least 10 AUT. It increases with Invocations skill and the difference
between player level and monster HD, but it takes a significant amount of
those factors to raise the duration significantly. Against a monster with
HD = XL, 27 invo should raise the duration from 10 AUT to 40 AUT.

This triggers passively so that some aspects of the 'combo' this god offers
don't require manual intervention. Plus, it'll be a cool event to aim for.

The ability only triggers at most once every 15 turns (150 AUT) so that
it can't be abused via repeated piety loss/gain.