Memoize several feature test functions.
These functions iterated over all branches looking for the queried
feature as the branch's entrance or exit feature.  But the answer for a
given query feature is never going to change.

Memoizing these feature tests is particularly important because they are
called extremely frequently.  This change results in an approximately
30% reduction in runtime for 1000 turns of the abyss walker bot, most of
which gprof attributed to calls to branch_iterator::operator->().

Also suggest a ctags option that recognizes uses of the new macro as
function definitions.