Remove various wands
Cold, fire, frost, magic darts, and invisibility. Wands
of lightning, draining, fireball, and flame are
now more common. Random Effects still chooses from the old wand
effects, primarily because I'd be sad if it could no longer turn
enemies invisible; that means it's now open to being changed
further, of course.

The primary reasoning here is that, as we've added new evocables
and other items, we haven't made any compensating removals; this
means that players inventories are more cluttered (which presents
annoying UI issues) and have an excess of tactical options. The
current set of wands is at least a decade old (!), which makes it
low-hanging fruit for pruning.

Darts, frost, fire & cold are all quite bland; flame covers the
former two just fine, and draining, lightning & fireball, while
not quite as strong as fire/cold, make a more distinctive trio.
(If the latter two get used a little more, that'd be a bonus.)
Invisibility is still available on potions, randarts, and as a
spell; it also doesn't have the ally synergy that the other buff
wands have.