Simplify and make permanent the removed wand checks.
If we defer to item_type_removed, we don't need to remove the code on a
major version bump, or to remember to update both locations the next
time a wand is removed.

Also, remove a cast-to-enum in a switch condition: all it did was give
devious compilers a theoretical justification for optimizing out the
"bugginess" case.