Make the captain's cutlass deal bonus damage when disarming
The formula for disarming success has been changed and it is now based
on target HD.

Enchantment has been improved from +5 to +7. (was +10 pre-disarm)

In the event of a disarm, 18-35 (average 26.5) non-elemental damage is
dealt to the target. This is akin to having a powerful flat damage
brand versus monsters wielding a weapon, but it can only trigger once
per monster. (twice for dual-wielders!)

This effect is considered unchivalric for TSO followers and will
cause penance.