Wands: draining -> acid, fireball -> iceblast
Replaces the wand of draining with the wand of acid, which fires a
corrosive bolt. This is currently roughly equally powerful (though
the damage type better), but may be buffed later to become some
kind of 'super wand'.

Without draining, there's now room to add an ice wand back (which
we'd wanted!), so fireball is out and iceblast is in. Partially
resistable cold (as throw icicle), quieter explosion than fireball.