Catch a few exceptions and die more precisely (PleasingFungus)
When calling "crawl -mapstat junk", we should print the error message
and exit normally, rather than crashing with "terminate called after
throwing ...".

When an old (very) obsolete ENTER_PORTAL_VAULT has a bad destination,
we do want to crash, but it's still nice to have the actual error
message.

Also, mark in the interface for depth_ranges::parse_depth_ranges that it
throws strings, for consistency with the level_range parsing functions.
I am unhappy with this situation for two reasons:

  * That's not what throw(foo) means in C++.  It actually means
    "I guarantee I won't throw any exceptions other than foo, and if
    I do so, call std::unexpected to abort".

  * We shouldn't be throwing raw strings anyway.

But in both cases, we do that all over the place.  Fixing the second
problem in particular was "literally impossible" in the learndb sense,
so has been deferred.  Enjoy the technical debt.