Remove Stoneskin※
It's an extremely generic, non-situational buff that it's wise,
but not particularly fun, to keep running at all times. It's not
particularly core to EE, so it's getting the axe. (I wouldn't mind
EE getting some replacement spell - but that's not critical.)

※It's still around for monsters (though I'm not sure any cast it),
and as an effect from the Helm card.