Simplify Song of Slaying
Rather than having an invisible, threat-based check to see whether
a given kill 'counts' for purpose of giving +slay, every kill now
gives +slay, up to +9 max.

The previous system was clearly intended to avoid abuse (towing
around weak enemies to farm slay), but that was *still* encouraged
by this system, just in an invisible & undocumented way. This
version might be too much of a buff, but that can be adjusted in
other ways (duration, slaycap, noisiness) as needed.

XXX: would be good to document the slay cap in the spell
description, somehow - I couldn't think of a non-🧀y phrasing.