Replace death cobs' hungering attack with AF_DRAIN_SPEED
On such a fast monster, AF_HUNGER tended to only allow for a fairly limited
response, since actually staying in combat and eating to counteract it was
pretty ineffective.

AF_DRAIN_SPEED keeps some of the intended behaviour of the hungering effect
(giving the player less turns in combat, due to plain slowness rather than
having to spend some turns eating), and should still be a relatively dangerous
effect in Zot.