Remove Condensation Shield
As with Stoneskin, it's not a well-designed effect for a spell, and IE can
handle the loss of it from the starting book with no problem. It's only used
very rarely for monsters, so rather than keep it around just for a couple of
cases, remove the effect entirely. The few monsters that had it get a couple
of extra points of EV (and the second deep elf knight spellbook loses Shroud,
as well).