Deduplicate and fix save-compat map feature rewriting.
When loading a save from before the removal of mangroves, we unmarshalled
malign gateways into grd correctly, but failed to shift the enum down in
map_knowledge.  Avoid these inconsistencies by moving all this logic,
as well as some other duplicated feature-rewriting logic, into a function
rewrite_feature.

This does theoretically affect the unmarshalling of you.transit_stairs
(it now undergoes some rewriting that it did not before), but that seems
like the correct thing to do.  This change would not make a difference
currently: the only check related to stairs that was added in this
codepath is to rewrite escape hatches when the player is in Labyrinth;
and a player transiting *into* Lab would have enter_lab, not an escape
hatch, as their transit stair.