Let Ukayaw's trigger trigger randomly if you stay at high enough piety
If your piety is high enough, each turn there's a chance of a trigger going
off; eventually the timer will reset and the trigger will happen again.
This meant to help resolve the issue of keeping piety too high w/ Ukayaw
in sustained fights. It's still beneficial to drop your piety with Grand
Finale so that you can keep more control over the timing of the triggers.

Both triggers have a aut_taken in aut_taken * 10 + timer_in_auts chance to
trigger on a given turn while in timeout.

The timers on both triggers have been increased by 100 AUT to partially
compensate for this boost.