Make a few exact holiness checks into flag checks.
Four changes:

  * Chaos attacks no longer try to poison dual-holiness undead.

  * Chaos attacks will try to drain/vamp all natural monsters, not
    just pure-natural.  This means Jessica can again be drained by a
    player's chaos weapon.

  * Crypt's _place_assorted_zombies can now place zombies of natural+
    monsters (hell hounds, for example).

  * All brainless natural creatures can be frenzied.  No extant monsters
    are affected, since currently all brainless natural monsters are
    pure natural.

The exact holiness checks remaining in the code are:

  * pure holies canot be placed by Lugonu's Corrupt, even though they
    do spawn "naturally" in the Abyss.

  * only pure holies get TSO as their god.

  * only pure undead are considered Crypt-native.

  * only pure demons trigger the DID_KILL_DEMON conduct.

The first three are unproblematic I think.  For the last, to allow all
relevant gods to get piety for such monsters, while avoiding double
piety, I suggest having getting (in mon-death) a list of all the
DID_KILL_*s for which this monster qualifies, then picking the one most
relevant to the player's god.  So it's more complicated than just
turning an == into an &.