Automatically respawn the ancestor
About 30-60 turns after it's destroyed, or 5-15 turns after it's
swallowed by an Abyss shift or similar.

The idea here is that the ancestor being around should be a default,
passive effect, rather than one you have to spend piety to maintain
(whenever it dies); we'll see how it works in practice.