Add Hepliaklqana death/swap effects
A one-time choice at 6*, between three provided options out of five
possible; an effect that triggers when your ancestor dies, or when
you swap with them with the Transference ability [added in later
commit. The effect triggers on your ancestor's prior location, so
either where they died or where you are now (if you swapped).

- Slow: Slows nearby enemies. Radius 1 on swap, 2 on death.
- Dispersal: As Dispersal. Power 40 on swap, 100 on death.
- Fog: Pops fog. Similar to a fog scroll; more powerful on death.
- Implode: As Gell's Gravitas. Higher power on death.
- Explosion: Damaging explosion. Doesn't harm you or the ancestor.
  Larger & more damage on death.

Goal: more customizability/personality for the ancestor, more
distinctiveness & tactical fun with the swap ability.

Problems: kind of complex and hard to explain; might be better if
it only triggered on swap...?