Hepliaklqana 5* ability: Transference
Swaps you with your ancestor if they're in LOS, triggering a
eathswap effect if you've chosen one. Takes a bit of piety.

Intended to be a fun tactical trick; problem is that allies have
monster AI, meaning they tend to charge at enemies, meaning the
positioning options are in practice limited. Will have to see how
well it plays.