Allow browsing char dump in-game with ?#
This command does not actually write out the dump to a file: in
principle, there is no reason it could not, but this way players have a
way to look at their current character info (action counts for example)
without overwriting the previous dump.

This required adjusting dump_params so that it owns its own text string
rather than a reference.  This does mean the text string is copied one
extra time in # and in ?#, unless the C++ return value optimization
kicks in.

In order to avoid misinterpreting "<" characters in the dump, add and
use the method formatted_scroller::add_raw_text().  It works like
add_text, except it does not attempt to parse colour strings.