Cut Summon Demon from deep elf high priests, sorcerers spell lists
Deep elf demonologists, elementalists, and 4.5/6 of mages already vaguely
prominently summon, so the common default-spell-filler of summon demon on
high priests and sorcerers seems to be rather extraneous for the monster group
(and they've got their own spells to work with). Casting weight evenly
distributed to a bit more smiting and malign offering on high priests,
partially slid onto corrosive bolt for sorcerers.

Sole vault that supposes either of them cast summon demon is untouched for
variety and it already not reliably having much summoners. Translations for
the sorcerer description are scrapped due to inaccuracy (and the current one
is mediocre anyway, as neither corrosion nor banishment are actually very
Hellish).