New monster: Kobold Cultist WIP
Another application of the sputtercasting mechanic (as seen on orb
spiders & salamander stormcallers), an early-game banishment
monster that needs to 'charge up' before actually setting off the
banishment. Numbers are all wrong, placement's not done, description
& speech lines aren't done, and I'm not sold on maintain_range;
it needs *some* mechanic to keep it alive long enough to do
anything, but this might not be it.