Replace Hunting Cry with Warning Cry
Hunting Cry's behaviour of alerting certain monsters was unclear as well as
being special-cased to work differently on different casters. The special
effect was also unlikely to be very significant, and generally didn't justify
it existing alongside Warning Cry (which is nice and simple and just makes a
bunch of noise).