Make water nymphs more fragile
They're the (or one of the) nastiest monsters in the nastiest lair
branch; it seems reasonable to tone them down a little. They'd be
fine more uncommon top-end enemies, but it seems more interesting
to have them appear regularly throughout the branch. The super
lethal waterstrike is their biggest distinctive trait, so instead
of nerfing that, make them a bit more fragile. (-20% HP and a bit
of AC.)