Farmer background.
Starts with 1* Fedhas piety, a scythe, a robe, fruit in place of
bread rations (but not meat rations), and three base skill points in
Polearms, Dodging, and Invocations.

Hopefully players will appreciate this contribution to the ecosystem.

(There is a pattern in the recommended species - see if you can deduce
it.)